[Decomp][Player] Cohesion split: extract item storage methods to PlayerItemStorage.cpp
